Update README.md
In this project, we built Ensemble models on Bank 'credit_score' data to predict default customers. Previously, we set up the logisticRegression model for this bank data and get 58% Gini. But this time we built Ensemble models again on the Bank 'credit_score' data and we managed to get a very significant Gini increase (74%).
Several models were used at the same time in this project. The Catboost model, which can operate in both categorical and numeric columns and the same time, we can count the XCBM model and one of the most important models, Stacking. Also, RandomForest, XGB and LogisticRegression models are used in the base part of the Stacking model. And hyperparameters too.